+++ src/Controller/AuthController.php
@@ -106,8 +106,11 @@ class AuthController extends AbstractActionController
         try {
             $response = $oauth2server->handleTokenRequest($oauth2request);
         } catch (ProblemExceptionInterface $ex) {
+            $status = $ex->getCode() ?: 401;
+            $status = $status >=400 && $status < 600 ? $status : 401;
+
             return new ApiProblemResponse(
-                new ApiProblem(401, $ex)
+                new ApiProblem($status, $ex)
             );
         }